Updated LMI to have an 'OS' module, fixed error of dynamic importing
Updated LMI to have OS module which allows for data to be collected about current os and able to retrieve the exact memory address of lucifer manager with the module. 
Error with dynamic modules fixed, as program running modules can be written to and edited and reloaded with "use <module> -R" for easy editing of modules for developing.